// front/src/interfaces/Business.ts
export default interface IBusiness {
    location_id?: any;
    altitude?: number;
    longitude?: number;
    email?: string;
    password?:string;
    name?: string;
    phone?: string;
    address?: string;
    status?:string;
    business_type_id?: number;
	business_activity_id?: number;
    placement?:string;
};